+++ lib/jxl/base/common.h
@@ -12,10 +12,15 @@
 #include <cstddef>
 #include <cstdint>
 #include <cstdio>
+#include <limits>
 #include <memory>
 #include <string>
 #include <type_traits>
 #include <vector>
+
+#if JXL_COMPILER_MSVC
+#include <intrin.h>
+#endif
 
 #include "lib/jxl/base/compiler_specific.h"
 
@@ -32,10 +37,36 @@ constexpr inline size_t RoundUpToBlockDim(size_t dim) {
   return (dim + 7) & ~static_cast<size_t>(7);
 }
 
-static inline bool JXL_MAYBE_UNUSED SafeAdd(const uint64_t a, const uint64_t b,
-                                            uint64_t& sum) {
+template <typename U,
+          class = typename std::enable_if<std::is_unsigned<U>::value>::type>
+static inline bool SafeAdd(const U a, const U b, U& sum) {
   sum = a + b;
   return sum >= a;  // no need to check b - either sum >= both or < both.
+}
+
+template <typename U,
+          class = typename std::enable_if<std::is_unsigned<U>::value>::type>
+static inline bool SafeMul(const U a, const U b, U& product) {
+  product = 0;
+  if (a == 0 || b == 0) return true;
+  if (b > (std::numeric_limits<U>::max() / a)) return false;
+  product = a * b;
+  return true;
+}
+
+static inline bool SubOverflow(const int32_t a, const int32_t b, int32_t& c) {
+  // Clang 3.8+ / GCC 5.1+
+#if JXL_COMPILER_GCC || JXL_COMPILER_CLANG
+  return __builtin_sub_overflow(a, b, &c);
+#elif JXL_COMPILER_MSVC >= 1937
+  return _sub_overflow_i32(/*carry*/ 0, a, b, &c);
+#else
+  uint32_t ua = static_cast<uint32_t>(a);
+  uint32_t ub = static_cast<uint32_t>(b);
+  uint32_t uc = ua - ub;
+  c = static_cast<int32_t>(uc);
+  return !!(((ua ^ ub) & (ua ^ uc)) >> 31);
+#endif
 }
 
 template <typename T1, typename T2>