+++ lib/jxl/render_pipeline/stage_epf.cc
@@ -53,8 +53,8 @@ JXL_INLINE Vec<DF> Weight(Vec<DF> sad, Vec<DF> inv_sigma, Vec<DF> thres) {
 class EPF0Stage : public RenderPipelineStage {
  public:
   EPF0Stage(LoopFilter lf, const ImageF& sigma)
-      : RenderPipelineStage(RenderPipelineStage::Settings::Symmetric(
-            /*shift=*/0, /*border=*/3)),
+      : RenderPipelineStage(
+            RenderPipelineStage::Settings::SymmetricBorderOnly(3)),
         lf_(std::move(lf)),
         sigma_(&sigma) {}
 
@@ -79,15 +79,18 @@ class EPF0Stage : public RenderPipelineStage {
   }
 
   Status ProcessRow(const RowInfo& input_rows, const RowInfo& output_rows,
-                    size_t xextra, size_t xsize, size_t xpos, size_t ypos,
-                    size_t thread_id) const final {
+                    size_t xextra_left, size_t xextra_right, size_t xsize,
+                    size_t xpos, size_t ypos, size_t thread_id) const final {
     DF df;
 
     using V = decltype(Zero(df));
     V t0, t1, t2, t3, t4, t5, t6, t7, t8, t9, tA, tB;  // NOLINT
     V* sads[12] = {&t0, &t1, &t2, &t3, &t4, &t5, &t6, &t7, &t8, &t9, &tA, &tB};
 
-    xextra = RoundUpTo(xextra, Lanes(df));
+    ptrdiff_t x_start =
+        -static_cast<ptrdiff_t>(RoundUpTo(xextra_left, Lanes(df)));
+    ptrdiff_t x_end = static_cast<ptrdiff_t>(xsize + xextra_right);
+
     const float* JXL_RESTRICT row_sigma =
         sigma_->Row(ypos / kBlockDim + kSigmaPadding);
 
@@ -110,14 +113,14 @@ class EPF0Stage : public RenderPipelineStage {
             ? sad_mul_border
             : sad_mul_center;
 
-    for (ptrdiff_t x = -xextra; x < static_cast<ptrdiff_t>(xsize + xextra);
-         x += Lanes(df)) {
+    for (ptrdiff_t x = x_start; x < x_end; x += Lanes(df)) {
       size_t bx = (x + xpos + kSigmaPadding * kBlockDim) / kBlockDim;
       size_t ix = (x + xpos) % kBlockDim;
 
       if (row_sigma[bx] < kMinSigma) {
         for (size_t c = 0; c < 3; c++) {
           auto px = Load(df, rows[c][3 + 0] + x);
+          // TODO(eustas): why unaligned?
           StoreU(px, df, GetOutputRow(output_rows, c, 0) + x);
         }
         continue;
@@ -168,6 +171,7 @@ class EPF0Stage : public RenderPipelineStage {
 #else
       auto inv_w = ApproximateReciprocal(w);
 #endif
+      // TODO(eustas): why unaligned?
       StoreU(Mul(X, inv_w), df, GetOutputRow(output_rows, 0, 0) + x);
       StoreU(Mul(Y, inv_w), df, GetOutputRow(output_rows, 1, 0) + x);
       StoreU(Mul(B, inv_w), df, GetOutputRow(output_rows, 2, 0) + x);
@@ -192,8 +196,8 @@ class EPF0Stage : public RenderPipelineStage {
 class EPF1Stage : public RenderPipelineStage {
  public:
   EPF1Stage(LoopFilter lf, const ImageF& sigma)
-      : RenderPipelineStage(RenderPipelineStage::Settings::Symmetric(
-            /*shift=*/0, /*border=*/2)),
+      : RenderPipelineStage(
+            RenderPipelineStage::Settings::SymmetricBorderOnly(2)),
         lf_(std::move(lf)),
         sigma_(&sigma) {}
 
@@ -218,10 +222,14 @@ class EPF1Stage : public RenderPipelineStage {
   }
 
   Status ProcessRow(const RowInfo& input_rows, const RowInfo& output_rows,
-                    size_t xextra, size_t xsize, size_t xpos, size_t ypos,
-                    size_t thread_id) const final {
+                    size_t xextra_left, size_t xextra_right, size_t xsize,
+                    size_t xpos, size_t ypos, size_t thread_id) const final {
     DF df;
-    xextra = RoundUpTo(xextra, Lanes(df));
+
+    ptrdiff_t x_start =
+        -static_cast<ptrdiff_t>(RoundUpTo(xextra_left, Lanes(df)));
+    ptrdiff_t x_end = static_cast<ptrdiff_t>(xsize + xextra_right);
+
     const float* JXL_RESTRICT row_sigma =
         sigma_->Row(ypos / kBlockDim + kSigmaPadding);
 
@@ -245,8 +253,8 @@ class EPF1Stage : public RenderPipelineStage {
             ? sad_mul_border
             : sad_mul_center;
 
-    for (ptrdiff_t x = -xextra; x < static_cast<ptrdiff_t>(xsize + xextra);
-         x += Lanes(df)) {
+    // TODO(eustas): cache GetOutputRow
+    for (ptrdiff_t x = x_start; x < x_end; x += Lanes(df)) {
       size_t bx = (x + xpos + kSigmaPadding * kBlockDim) / kBlockDim;
       size_t ix = (x + xpos) % kBlockDim;
 
@@ -285,6 +293,7 @@ class EPF1Stage : public RenderPipelineStage {
         sad1c = Add(sad1c, AbsDiff(p02, p12));  // SAD 1, 2
         sad0c = Add(sad0c, AbsDiff(p11, p12));  // SAD 2, 1
 
+        // TODO(eustas): why unaligned?
         const auto p22 = LoadU(df, rows[c][2 + 0] + x);
         t = AbsDiff(p12, p22);
         sad1c = Add(sad1c, t);  // SAD 1, 2
@@ -373,8 +382,8 @@ class EPF1Stage : public RenderPipelineStage {
 class EPF2Stage : public RenderPipelineStage {
  public:
   EPF2Stage(LoopFilter lf, const ImageF& sigma)
-      : RenderPipelineStage(RenderPipelineStage::Settings::Symmetric(
-            /*shift=*/0, /*border=*/1)),
+      : RenderPipelineStage(
+            RenderPipelineStage::Settings::SymmetricBorderOnly(1)),
         lf_(std::move(lf)),
         sigma_(&sigma) {}
 
@@ -404,10 +413,14 @@ class EPF2Stage : public RenderPipelineStage {
   }
 
   Status ProcessRow(const RowInfo& input_rows, const RowInfo& output_rows,
-                    size_t xextra, size_t xsize, size_t xpos, size_t ypos,
-                    size_t thread_id) const final {
+                    size_t xextra_left, size_t xextra_right, size_t xsize,
+                    size_t xpos, size_t ypos, size_t thread_id) const final {
     DF df;
-    xextra = RoundUpTo(xextra, Lanes(df));
+
+    ptrdiff_t x_start =
+        -static_cast<ptrdiff_t>(RoundUpTo(xextra_left, Lanes(df)));
+    ptrdiff_t x_end = static_cast<ptrdiff_t>(xsize + xextra_right);
+
     const float* JXL_RESTRICT row_sigma =
         sigma_->Row(ypos / kBlockDim + kSigmaPadding);
 
@@ -431,8 +444,8 @@ class EPF2Stage : public RenderPipelineStage {
             ? sad_mul_border
             : sad_mul_center;
 
-    for (ptrdiff_t x = -xextra; x < static_cast<ptrdiff_t>(xsize + xextra);
-         x += Lanes(df)) {
+    // TODO(eustas): cache GetOutputRow
+    for (ptrdiff_t x = x_start; x < x_end; x += Lanes(df)) {
       size_t bx = (x + xpos + kSigmaPadding * kBlockDim) / kBlockDim;
       size_t ix = (x + xpos) % kBlockDim;