+++ lib/jxl/render_pipeline/stage_patches.cc
@@ -26,17 +26,16 @@ class PatchDictionaryStage : public RenderPipelineStage {
         extra_channel_info_(extra_channel_info) {}
 
   Status ProcessRow(const RowInfo& input_rows, const RowInfo& output_rows,
-                    size_t xextra, size_t xsize, size_t xpos, size_t ypos,
-                    size_t thread_id) const final {
-    JXL_ENSURE(xpos == 0 || xpos >= xextra);
-    size_t x0 = xpos ? xpos - xextra : 0;
+                    size_t xextra_left, size_t xextra_right, size_t xsize,
+                    size_t xpos, size_t ypos, size_t thread_id) const final {
     size_t num_channels = 3 + extra_channel_info_->size();
     std::vector<float*> row_ptrs(num_channels);
     for (size_t i = 0; i < num_channels; i++) {
-      row_ptrs[i] = GetInputRow(input_rows, i, 0) + x0 - xpos;
+      row_ptrs[i] = GetInputRow(input_rows, i, 0) - xextra_left;
     }
-    return patches_.AddOneRow(row_ptrs.data(), ypos, x0,
-                              xsize + xextra + xpos - x0, *extra_channel_info_);
+    return patches_.AddOneRow(row_ptrs.data(), ypos, xpos - xextra_left,
+                              xsize + xextra_left + xextra_right,
+                              *extra_channel_info_);
   }
 
   RenderPipelineChannelMode GetChannelMode(size_t c) const final {