+++ lib/jxl/render_pipeline/stage_to_linear.cc
@@ -115,21 +115,22 @@ class ToLinearStage : public RenderPipelineStage {
       : RenderPipelineStage(RenderPipelineStage::Settings()), valid_(false) {}
 
   Status ProcessRow(const RowInfo& input_rows, const RowInfo& output_rows,
-                    size_t xextra, size_t xsize, size_t xpos, size_t ypos,
-                    size_t thread_id) const final {
+                    size_t xextra_left, size_t xextra_right, size_t xsize,
+                    size_t xpos, size_t ypos, size_t thread_id) const final {
     const HWY_FULL(float) d;
-    const size_t xsize_v = RoundUpTo(xsize, Lanes(d));
+    JXL_ENSURE(xextra_left == 0 && xextra_right == 0);
+    size_t x_span_tail = RoundUpTo(xsize, Lanes(d)) - xsize;
     float* JXL_RESTRICT row0 = GetInputRow(input_rows, 0, 0);
     float* JXL_RESTRICT row1 = GetInputRow(input_rows, 1, 0);
     float* JXL_RESTRICT row2 = GetInputRow(input_rows, 2, 0);
     // All calculations are lane-wise, still some might require
     // value-dependent behaviour (e.g. NearestInt). Temporary unpoison last
     // vector tail.
-    msan::UnpoisonMemory(row0 + xsize, sizeof(float) * (xsize_v - xsize));
-    msan::UnpoisonMemory(row1 + xsize, sizeof(float) * (xsize_v - xsize));
-    msan::UnpoisonMemory(row2 + xsize, sizeof(float) * (xsize_v - xsize));
-    for (ptrdiff_t x = -xextra; x < static_cast<ptrdiff_t>(xsize + xextra);
-         x += Lanes(d)) {
+    msan::UnpoisonMemory(row0 + xsize, sizeof(float) * x_span_tail);
+    msan::UnpoisonMemory(row1 + xsize, sizeof(float) * x_span_tail);
+    msan::UnpoisonMemory(row2 + xsize, sizeof(float) * x_span_tail);
+    // TODO(eustas): why unaligned load/store?
+    for (size_t x = 0; x < xsize; x += Lanes(d)) {
       auto r = LoadU(d, row0 + x);
       auto g = LoadU(d, row1 + x);
       auto b = LoadU(d, row2 + x);
@@ -138,9 +139,9 @@ class ToLinearStage : public RenderPipelineStage {
       StoreU(g, d, row1 + x);
       StoreU(b, d, row2 + x);
     }
-    msan::PoisonMemory(row0 + xsize, sizeof(float) * (xsize_v - xsize));
-    msan::PoisonMemory(row1 + xsize, sizeof(float) * (xsize_v - xsize));
-    msan::PoisonMemory(row2 + xsize, sizeof(float) * (xsize_v - xsize));
+    msan::PoisonMemory(row0 + xsize, sizeof(float) * x_span_tail);
+    msan::PoisonMemory(row1 + xsize, sizeof(float) * x_span_tail);
+    msan::PoisonMemory(row2 + xsize, sizeof(float) * x_span_tail);
     return true;
   }