+++ tools/benchmark/benchmark_codec_jpeg.cc
@@ -19,25 +19,18 @@
 #include <vector>
 
 #include "lib/extras/codec.h"
-#include "lib/extras/enc/encode.h"
-#include "lib/jxl/base/status.h"
-#include "tools/benchmark/benchmark_args.h"
-#include "tools/benchmark/benchmark_codec.h"
-#include "tools/speed_stats.h"
-
-#if JPEGXL_ENABLE_JPEGLI
-#include "lib/extras/dec/jpegli.h"
-#endif
 #include "lib/extras/dec/jpg.h"
-#if JPEGXL_ENABLE_JPEGLI
-#include "lib/extras/enc/jpegli.h"
-#endif
+#include "lib/extras/enc/encode.h"
 #include "lib/extras/enc/jpg.h"
 #include "lib/extras/packed_image.h"
 #include "lib/extras/time.h"
 #include "lib/jxl/base/span.h"
+#include "lib/jxl/base/status.h"
+#include "tools/benchmark/benchmark_args.h"
+#include "tools/benchmark/benchmark_codec.h"
 #include "tools/benchmark/benchmark_utils.h"
 #include "tools/file_io.h"
+#include "tools/speed_stats.h"
 #include "tools/thread_pool_internal.h"
 
 namespace jpegxl {
@@ -109,12 +102,6 @@ class JPEGCodec : public ImageCodec {
       jpeg_encoder_ = param;
       return true;
     }
-#if JPEGXL_ENABLE_JPEGLI
-    if (param == "enc-jpegli") {
-      jpeg_encoder_ = "jpegli";
-      return true;
-    }
-#endif
     if (param.compare(0, 3, "yuv") == 0) {
       chroma_subsampling_ = param.substr(3);
       return true;
@@ -144,28 +131,6 @@ class JPEGCodec : public ImageCodec {
       enable_adaptive_quant_ = false;
       return true;
     }
-#if JPEGXL_ENABLE_JPEGLI
-    if (param == "xyb") {
-      xyb_mode_ = true;
-      return true;
-    }
-    if (param == "std") {
-      use_std_tables_ = true;
-      return true;
-    }
-    if (param == "dec-jpegli") {
-      jpeg_decoder_ = "jpegli";
-      return true;
-    }
-    if (param.substr(0, 2) == "bd") {
-      bitdepth_ = strtol(param.substr(2).c_str(), nullptr, 10);
-      return true;
-    }
-    if (param.substr(0, 6) == "cquant") {
-      num_colors_ = strtol(param.substr(6).c_str(), nullptr, 10);
-      return true;
-    }
-#endif
     return false;
   }
 
@@ -216,94 +181,54 @@ class JPEGCodec : public ImageCodec {
     }
 
     double elapsed = 0.0;
-    if (jpeg_encoder_ == "jpegli") {
-#if JPEGXL_ENABLE_JPEGLI
-      jxl::extras::JpegSettings settings;
-      settings.xyb = xyb_mode_;
-      if (!xyb_mode_) {
-        settings.use_std_quant_tables = use_std_tables_;
-      }
-      if (enc_quality_set_) {
-        settings.quality = q_target_;
-      } else {
-        settings.distance = butteraugli_target_;
-      }
-      if (progressive_id_ >= 0) {
-        settings.progressive_level = progressive_id_;
-      }
-      if (psnr_target_ > 0) {
-        settings.psnr_target = psnr_target_;
-      }
-      if (jpegargs->search_tolerance > 0) {
-        settings.search_tolerance = 0.01f * jpegargs->search_tolerance;
-      }
-      if (jpegargs->search_d_min > 0) {
-        settings.min_distance = jpegargs->search_d_min;
-      }
-      if (jpegargs->search_d_max > 0) {
-        settings.max_distance = jpegargs->search_d_max;
-      }
-      settings.chroma_subsampling = chroma_subsampling_;
-      settings.use_adaptive_quantization = enable_adaptive_quant_;
-      settings.libjpeg_quality = libjpeg_quality_;
-      settings.libjpeg_chroma_subsampling = libjpeg_chroma_subsampling_;
-      settings.optimize_coding = !fix_codes_;
-      const double start = jxl::Now();
-      JXL_RETURN_IF_ERROR(
-          jxl::extras::EncodeJpeg(ppf, settings, pool, compressed));
-      const double end = jxl::Now();
-      elapsed = end - start;
-#endif
-    } else {
-      jxl::extras::EncodedImage encoded;
-      std::unique_ptr<jxl::extras::Encoder> encoder =
-          jxl::extras::GetJPEGEncoder();
-      if (!encoder) {
-        fprintf(stderr, "libjpeg codec is not supported\n");
-        return false;
-      }
-      std::ostringstream os;
-      os << static_cast<int>(std::round(q_target_));
-      encoder->SetOption("q", os.str());
-      encoder->SetOption("jpeg_encoder", jpeg_encoder_);
-      if (!chroma_subsampling_.empty()) {
-        encoder->SetOption("chroma_subsampling", chroma_subsampling_);
-      }
-      if (progressive_id_ >= 0) {
-        encoder->SetOption("progressive", std::to_string(progressive_id_));
-      }
-      if (libjpeg_quality_ > 0) {
-        encoder->SetOption("libjpeg_quality", std::to_string(libjpeg_quality_));
-      }
-      if (!libjpeg_chroma_subsampling_.empty()) {
-        encoder->SetOption("libjpeg_chroma_subsampling",
-                           libjpeg_chroma_subsampling_);
-      }
-      if (fix_codes_) {
-        encoder->SetOption("optimize", "OFF");
-      }
-      if (!enable_adaptive_quant_) {
-        encoder->SetOption("adaptive_q", "OFF");
-      }
-      if (psnr_target_ > 0) {
-        encoder->SetOption("psnr", std::to_string(psnr_target_));
-      }
-      if (!jpegargs->base_quant_fn.empty()) {
-        encoder->SetOption("base_quant_fn", jpegargs->base_quant_fn);
-      }
-      SET_ENCODER_ARG(search_q_start);
-      SET_ENCODER_ARG(search_q_min);
-      SET_ENCODER_ARG(search_q_max);
-      SET_ENCODER_ARG(search_q_precision);
-      SET_ENCODER_ARG(search_tolerance);
-      SET_ENCODER_ARG(search_first_iter_slope);
-      SET_ENCODER_ARG(search_max_iters);
-      const double start = jxl::Now();
-      JXL_RETURN_IF_ERROR(encoder->Encode(ppf, &encoded, pool));
-      const double end = jxl::Now();
-      elapsed = end - start;
-      *compressed = encoded.bitstreams.back();
+    jxl::extras::EncodedImage encoded;
+    std::unique_ptr<jxl::extras::Encoder> encoder =
+        jxl::extras::GetJPEGEncoder();
+    if (!encoder) {
+      fprintf(stderr, "libjpeg codec is not supported\n");
+      return false;
     }
+    std::ostringstream os;
+    os << static_cast<int>(std::round(q_target_));
+    encoder->SetOption("q", os.str());
+    encoder->SetOption("jpeg_encoder", jpeg_encoder_);
+    if (!chroma_subsampling_.empty()) {
+      encoder->SetOption("chroma_subsampling", chroma_subsampling_);
+    }
+    if (progressive_id_ >= 0) {
+      encoder->SetOption("progressive", std::to_string(progressive_id_));
+    }
+    if (libjpeg_quality_ > 0) {
+      encoder->SetOption("libjpeg_quality", std::to_string(libjpeg_quality_));
+    }
+    if (!libjpeg_chroma_subsampling_.empty()) {
+      encoder->SetOption("libjpeg_chroma_subsampling",
+                         libjpeg_chroma_subsampling_);
+    }
+    if (fix_codes_) {
+      encoder->SetOption("optimize", "OFF");
+    }
+    if (!enable_adaptive_quant_) {
+      encoder->SetOption("adaptive_q", "OFF");
+    }
+    if (psnr_target_ > 0) {
+      encoder->SetOption("psnr", std::to_string(psnr_target_));
+    }
+    if (!jpegargs->base_quant_fn.empty()) {
+      encoder->SetOption("base_quant_fn", jpegargs->base_quant_fn);
+    }
+    SET_ENCODER_ARG(search_q_start);
+    SET_ENCODER_ARG(search_q_min);
+    SET_ENCODER_ARG(search_q_max);
+    SET_ENCODER_ARG(search_q_precision);
+    SET_ENCODER_ARG(search_tolerance);
+    SET_ENCODER_ARG(search_first_iter_slope);
+    SET_ENCODER_ARG(search_max_iters);
+    const double start = jxl::Now();
+    JXL_RETURN_IF_ERROR(encoder->Encode(ppf, &encoded, pool));
+    const double end = jxl::Now();
+    elapsed = end - start;
+    *compressed = encoded.bitstreams.back();
     speed_stats->NotifyElapsed(elapsed);
     return true;
   }
@@ -312,30 +237,14 @@ class JPEGCodec : public ImageCodec {
                     const Span<const uint8_t> compressed, ThreadPool* pool,
                     jxl::extras::PackedPixelFile* ppf,
                     jpegxl::tools::SpeedStats* speed_stats) override {
-    if (jpeg_decoder_ == "jpegli") {
-#if JPEGXL_ENABLE_JPEGLI
-      std::vector<uint8_t> jpeg_bytes(compressed.data(),
-                                      compressed.data() + compressed.size());
-      const double start = jxl::Now();
-      jxl::extras::JpegDecompressParams dparams;
-      dparams.output_data_type =
-          bitdepth_ > 8 ? JXL_TYPE_UINT16 : JXL_TYPE_UINT8;
-      dparams.num_colors = num_colors_;
-      JXL_RETURN_IF_ERROR(
-          jxl::extras::DecodeJpeg(jpeg_bytes, dparams, pool, ppf));
-      const double end = jxl::Now();
-      speed_stats->NotifyElapsed(end - start);
-#endif
-    } else {
-      const double start = jxl::Now();
-      jxl::extras::JPGDecompressParams dparams;
-      dparams.num_colors = num_colors_;
-      JXL_RETURN_IF_ERROR(
-          jxl::extras::DecodeImageJPG(compressed, jxl::extras::ColorHints(),
-                                      ppf, /*constraints=*/nullptr, &dparams));
-      const double end = jxl::Now();
-      speed_stats->NotifyElapsed(end - start);
-    }
+    const double start = jxl::Now();
+    jxl::extras::JPGDecompressParams dparams;
+    dparams.num_colors = num_colors_;
+    JXL_RETURN_IF_ERROR(
+        jxl::extras::DecodeImageJPG(compressed, jxl::extras::ColorHints(), ppf,
+                                    /*constraints=*/nullptr, &dparams));
+    const double end = jxl::Now();
+    speed_stats->NotifyElapsed(end - start);
     return true;
   }
 
@@ -349,17 +258,10 @@ class JPEGCodec : public ImageCodec {
   bool enc_quality_set_ = false;
   int libjpeg_quality_ = 0;
   std::string libjpeg_chroma_subsampling_;
-#if JPEGXL_ENABLE_JPEGLI
-  bool xyb_mode_ = false;
-  bool use_std_tables_ = false;
-#endif
   bool enable_adaptive_quant_ = true;
   // JPEG decoder and its parameters
   std::string jpeg_decoder_ = "libjpeg";
   int num_colors_ = 0;
-#if JPEGXL_ENABLE_JPEGLI
-  size_t bitdepth_ = 8;
-#endif
 };
 
 ImageCodec* CreateNewJPEGCodec(const BenchmarkArgs& args) {